version-based pickup listener
