+++ IF/src/main/java/com/github/stefvanschie/inventoryframework/gui/listener/AbstractGuiListener.java
@@ -1,0 +1,36 @@
+package com.github.stefvanschie.inventoryframework.gui.listener;
+
+import com.github.stefvanschie.inventoryframework.gui.type.util.Gui;
+import org.bukkit.event.Listener;
+import org.bukkit.inventory.Inventory;
+import org.bukkit.inventory.InventoryHolder;
+import org.jetbrains.annotations.Contract;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
+
+public abstract class AbstractGuiListener implements Listener {
+    /**
+     * Gets the gui from the inventory or null if the inventory isn't a gui
+     *
+     * @param inventory the inventory to get the gui from
+     * @return the gui or null if the inventory doesn't have a gui
+     * @since 0.8.1
+     */
+    @Nullable
+    @Contract(pure = true)
+    protected final Gui getGui(@NotNull Inventory inventory) {
+        Gui gui = Gui.getGui(inventory);
+
+        if (gui != null) {
+            return gui;
+        }
+
+        InventoryHolder holder = inventory.getHolder();
+
+        if (holder instanceof Gui) {
+            return (Gui) holder;
+        }
+
+        return null;
+    }
+}
