+++ IF/src/main/java/com/github/stefvanschie/inventoryframework/gui/listener/OldPickupListener.java
@@ -1,0 +1,40 @@
+package com.github.stefvanschie.inventoryframework.gui.listener;
+
+import com.github.stefvanschie.inventoryframework.gui.type.util.Gui;
+import org.bukkit.event.EventHandler;
+import org.bukkit.event.EventPriority;
+import org.bukkit.event.player.PlayerPickupItemEvent;
+import org.bukkit.inventory.ItemStack;
+import org.jetbrains.annotations.NotNull;
+
+public class OldPickupListener extends AbstractGuiListener {
+    /**
+     * Handles users picking up items while their bottom inventory is in use.
+     *
+     * @param event the event fired when an entity picks up an item
+     * @since 0.6.1
+     */
+    @SuppressWarnings("deprecation")
+    @EventHandler(ignoreCancelled = true, priority = EventPriority.MONITOR)
+    public void onPlayerPickupItem(@NotNull PlayerPickupItemEvent event) {
+        Gui gui = getGui(event.getPlayer().getOpenInventory().getTopInventory());
+
+        if (gui == null || !gui.isPlayerInventoryUsed()) {
+            return;
+        }
+
+        int leftOver = gui.getHumanEntityCache().add(event.getPlayer(), event.getItem().getItemStack());
+
+        if (leftOver == 0) {
+            event.getItem().remove();
+        } else {
+            ItemStack itemStack = event.getItem().getItemStack();
+
+            itemStack.setAmount(leftOver);
+
+            event.getItem().setItemStack(itemStack);
+        }
+
+        event.setCancelled(true);
+    }
+}
